requirements.txt
asgiref==3.4.1
attrs==21.2.0
clize==4.2.0
Django==3.2.7
docutils==0.17.1
od==1.0
python-dateutil==2.8.2
python-gantt==0.6.0
pytz==2021.1
sigtools==2.0.2
six==1.16.0
sqlparse==0.4.1
svgwrite==1.4.1
typing-extensions==3.10.0.2
